Ajout du DTO et du controller des RDV pour les tests

+++ src/main/java/com/example/lexora/avocat/Avocat.java
@@ -34,11 +34,11 @@ public class Avocat extends User {
     @Column(name = "specialite")
     private List<String> specialite = new ArrayList<>();
 
-    private Integer note;
+    private Float note;
 
     private boolean isVerified;
 
-    private int honoraire;
+    private Integer honoraire;
 
     private Integer annee;
